+++ js/react/es5/DataFormsJS.js
@@ -1,5 +1,5 @@
 // @link https://www.dataformsjs.com
-// @version 4.5.2
+// @version 4.5.3
 // @author Conrad Sollitt (https://conradsollitt.com)
 // @license MIT
 "use strict";